# prepare perf: dont eagerly allocate result column name strings